Refactor (apps/pet-shop): Move `PSEmptinessComponent`
