+++ apps/pet-shop/src/app/integrator/ui-kit/ps-emptiness/emptiness.component.sass
